* master: (116 commits)
  Tidying up the changes.  DHB
  And more fixes from last change.  DHB
  Fix syntax typo from last change.  DHB
  Optional 1960 uv computations added, no change in default behavior.   Improved comments.  DHB
  * ExpandingRingsDemo:
  * PsychPortAudio: Improve scheduling of audio start in audio schedules.
  * PsychHID: Improve shutdown code for keyboard thread in KeyboardQueue   implementation. We pthread_join() the thread before releasing mutexes   etc. so we can't get into race-conditions between the shutting down thread   and the main thread.
  * Rebuild almost all Mex files for Matlab on OS/X PowerPC.
  * ExpandingRingsDemo.m: Add new optional shader for drawing   expanding sinusoidal color gratings, instead of rings with   hard edges.
  * PsychImaging: Recognize ATI on Gallium and allow to enable   natie 10 bpc framebuffer there as well.
  * Screen/GStreamer:
  Screen: Minor this and that...
  * Screen GPU detection code: Detect ATI/AMD GPU's when running   under Mesa+Gallium correctly, so their capabilities get detected.
  * Screen/GStreamer videocapture: Flexible codec settings.
  * Screen: Fix reporting of alpha-channel bit depths when           native hardware framebuffer > 8 bpc is enabled.
  * Screen/GStreamer vidcap:
  * Screen/GStreamer vidcap engine:
  Screen: Videocapture engines.
  * Screen/GStreamer videocapture engine:
  * Screen/GStreamer - Videocapture engine:
  ...

Conflicts ("resolved" using  debian/cleanup.sh script)
	PsychSourceGL/Cohorts/PsychtoolboxOSXKernelDriver/PsychGraphicsCardRegisterSpecs.h
	PsychSourceGL/Cohorts/PsychtoolboxOSXKernelDriver/PsychtoolboxKernelDriver.cpp
	PsychSourceGL/Cohorts/PsychtoolboxOSXKernelDriver/PsychtoolboxKernelDriver.h
	PsychSourceGL/Projects/MacOSX/PsychToolbox/PsychToolbox.xcodeproj/project.pbxproj
	PsychSourceGL/Source/OSX/Base/PsychTimeGlue.c
	PsychSourceGL/Source/OSX/Base/PsychTimeGlue.h
	PsychSourceGL/Source/OSX/EthernetAddress/GetEthernetAddress.c
	PsychSourceGL/Source/OSX/Fonts/PsychFontGlue.c
	PsychSourceGL/Source/OSX/Screen/PsychScreenGlue.c
	PsychSourceGL/Source/OSX/Screen/PsychWindowGlue.c
	PsychSourceGL/Source/OSX/Screen/PsychWindowTextGlue.c
	PsychSourceGL/Source/Windows/Base/PsychTimeGlue.h
	Psychtoolbox/PsychAlpha/PsychtoolboxKernelDriver.kext.zip
	Psychtoolbox/PsychBasic/DoNothing.mexmac
	Psychtoolbox/PsychBasic/Eyelink.mexglx
	Psychtoolbox/PsychBasic/FontInfo.mexmac
	Psychtoolbox/PsychBasic/Gestalt.mexmac
	Psychtoolbox/PsychBasic/GetBusTicks.mexmac
	Psychtoolbox/PsychBasic/GetPID.mexmac
	Psychtoolbox/PsychBasic/GetSecs.mexglx
	Psychtoolbox/PsychBasic/GetSecs.mexmac
	Psychtoolbox/PsychBasic/GetSecsTick.mexmac
	Psychtoolbox/PsychBasic/GetTicks.mexmac
	Psychtoolbox/PsychBasic/IOPort.mexglx
	Psychtoolbox/PsychBasic/IOPort.mexmac
	Psychtoolbox/PsychBasic/MachAbsoluteTimeClockFrequency.mexmac
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/GetSecs.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/IOPort.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/PsychPortAudio.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/Screen.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/WaitSecs.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR11/moglcore.dll
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/GetSecs.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/IOPort.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/PsychKinectCore.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/PsychPortAudio.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/Screen.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/WaitSecs.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/moalcore.mexw32
	Psychtoolbox/PsychBasic/MatlabWindowsFilesR2007a/moglcore.mexw32
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/Eyelink.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/GetSecs.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/IOPort.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/PsychKinectCore.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/PsychPortAudio.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/Screen.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/WaitSecs.mex
	Psychtoolbox/PsychBasic/Octave3LinuxFiles/moglcore.mex
	Psychtoolbox/PsychBasic/Octave3OSXFiles/PsychHID.mex
	Psychtoolbox/PsychBasic/Octave3OSXFiles/PsychKinectCore.mex
	Psychtoolbox/PsychBasic/Octave3OSXFiles/PsychPortAudio.mex
	Psychtoolbox/PsychBasic/Octave3OSXFiles/Screen.mex
	Psychtoolbox/PsychBasic/Octave3OSXFiles/moglcore.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/Eyelink.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/GetSecs.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/IOPort.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/PsychKinectCore.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/PsychPortAudio.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/Screen.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/WaitSecs.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/moglcore.mex
	Psychtoolbox/PsychBasic/Octave3WindowsFiles/pnet.mex
	Psychtoolbox/PsychBasic/PsychKinectCore.mexglx
	Psychtoolbox/PsychBasic/PsychKinectCore.mexmaci
	Psychtoolbox/PsychBasic/PsychPortAudio.mexglx
	Psychtoolbox/PsychBasic/PsychPortAudio.mexmac
	Psychtoolbox/PsychBasic/PsychPortAudio.mexmaci
	Psychtoolbox/PsychBasic/Screen.mexglx
	Psychtoolbox/PsychBasic/Screen.mexmac
	Psychtoolbox/PsychBasic/Screen.mexmaci
	Psychtoolbox/PsychBasic/WaitSecs.mexglx
	Psychtoolbox/PsychBasic/WaitSecs.mexmac
	Psychtoolbox/PsychDemos/ARToolkitDemo.m
	Psychtoolbox/PsychHardware/PsychHID.mexmac
	Psychtoolbox/PsychObsolete/GetBusFrequencyMex.mexmac
	Psychtoolbox/PsychObsolete/GetSecsMex.mexmac
	Psychtoolbox/PsychObsolete/MachPriority.mexmac
	Psychtoolbox/PsychOpenGL/MOGL/core/moglcore.mexglx
	Psychtoolbox/PsychOpenGL/MOGL/core/moglcore.mexmac
	Psychtoolbox/PsychOpenGL/MOGL/core/moglcore.mexmaci
	Psychtoolbox/PsychPriority/MachGetPriorityMex.mexmac
	Psychtoolbox/PsychPriority/MachSetPriorityMex.mexmac
	Psychtoolbox/PsychPriority/Rush.mexmac
	Psychtoolbox/PsychSound/MOAL/core/moalcore.mexmaci

+++ PsychSourceGL/Source/Common/Screen/WindowBank.h
@@ -234,7 +234,7 @@ typedef struct _PsychWindowRecordType_{
 	int					screenNumber;   // kPsychUnaffiliated is -1 and means the offscreen window is unaffiliated.
 	PsychWindowIndexType                    windowIndex;
 	void					*surface; 
-	int					surfaceSizeBytes;	// Estimate of used system memory in bytes. Only used for accounting and debugging output.
+	size_t					surfaceSizeBytes;	// Estimate of used system memory in bytes. Only used for accounting and debugging output.
 	PsychRectType              rect;		// Effective rectangle of window -- Normalized to always have top-left corner in (0,0)!
 	PsychRectType			globalrect;		// Same as rect, but not normalized -- Contains real window location in global desktop reference frame.
 	psych_bool					isValid;		//between when we allocate the record and when we fill in values.
@@ -299,7 +299,7 @@ typedef struct _PsychWindowRecordType_{
 		double									postflip_vbltimestamp;	// Optional timestamp taken after flip completion via PsychGetVBLTimeAndCount();
 		double									osbuiltin_swaptime;		// Optional timestamp of swap completion computed via PsychOSGetSwapCompletionTimestamp();
 		double									gpuRenderTime;			// GPU time spent on rendering. Only returned if a query object is successfully generated.
-		GLint									gpuRenderTimeQuery;		// Handle to the GPU time query object. 0 if none assigned.
+		GLuint									gpuRenderTimeQuery;		// Handle to the GPU time query object. 0 if none assigned.
 		psych_int64								reference_ust;			// UST reference timestamp of vblank with count reference_msc from OpenML. (Optional)
 		psych_int64								reference_msc;			// MSC reference vblank count from OpenML. (Optional)
 		psych_int64								reference_sbc;			// SBC reference swapbuffers count from OpenML. (Optional)
